Remove duplicate paths
With multiple files given as the input, it is
possible we get the same path more than once.
Under the default options this would not be a problem
because duplicate paths to the same file (by inode id)
are removed anyways in hardlink pruning step.
However, the user may deliberately
turn off by-id deduplication with -H option and in this
case duplicate paths would be treated as duplicate files.