Move deduplicating paths to the group-by-ids stage
Duplicate paths were removed at the end of
the group-by-size stage.
With millions of paths, this operation was costly
enough to make the progress pause noticeable.

In this commit path deduplicating was moved to
the later by-file-id deduplicating stage. This
way the progress is tracked nicely and, as a bonus,
we additionally get code simplification.