Update rayon to 1.4.0. No unsafe from pools.rs.
A tiny change in Rayon scope signature allowed
to get rid of unsafe.

Before:

    pub fn scope<'scope, OP, R>(&self, op: OP) -> R
    where
        OP: FnOnce(&Scope<'scope>) -> R + 'scope + Send,
        R: Send

After:

    pub fn scope<'scope, OP, R>(&self, op: OP) -> R
    where
        OP: FnOnce(&Scope<'scope>) -> R + Send,
        R: Send

This change allows 'scope to be inferred as a wider
lifetime than the lambda, hence all scopes
created in pools.rs get the same lifetime and can be
added to the same vector without unsafe tricks.