Scan dirs recursively by default
Option -R has been removed.
To get non-recursive scans, use --depth 0.

This change is a result of observation that most of the time,
users want to find duplicates in the whole subtree.
Non recursive search seems to be an edge case.
The interface should be optimized for the most common use case.