Implement `fclones dedupe` command
Deduplicate the underlying storage of files without
making visible changes to the files themselves.

If a deduplicated file is later written, the affected
blocks will first be copied (copy on write, or COW).

Currently only supported by the filesystems btrfs and
XFS on Linux, and APFS on macOS.

Will not fall back to another method on failure.