Fix handling of non-unicode and control chars in paths
Paths stored in reports are now in STFU-8 format,
which escapes non-UTF8, special and control characters,
so that text files produced by fclones are always valid UTF-8.
Additionally, newlines won't mess up the report file format.

A side effect of this change is that on Windows, the backslash
character ('\') used as a directory delimiter is
encoded as a double backslash '\\'.

Fixes #102