Don't hash files with same inode identifiers more than once
This optimization speeds up processing if the file tree contains
symbolic or hard links. Symbolic and hard links to the same data
will be hashed at most once. The algorithm is also smart enough
to not hash the data at all, if it determines that no other (unlinked)
matches are possible.

Fixes #139.